Limit CloudWatch permissions for Lambda function
This is much closer to what Lambda creates when you build a new function
in the web UI. The only missing piece is that the IAM role it creates
normally limits the function to log streams under its own special
namespace. Here, that causes a circular dependency between the role and
the function.